change timestamp to uint64 to fix #516

+++ platforms/leap/leap_motion_driver_test.go
@@ -89,7 +89,7 @@ func TestLeapMotionDriverParser(t *testing.T) {
 		t.Errorf("ParseFrame incorrectly parsed frame")
 	}
 
-	gobottest.Assert(t, parsedFrame.Timestamp, 4729292670)
+	gobottest.Assert(t, parsedFrame.Timestamp, uint64(4729292670))
 	gobottest.Assert(t, parsedFrame.Hands[0].X(), 117.546)
 	gobottest.Assert(t, parsedFrame.Hands[0].Y(), 236.007)
 	gobottest.Assert(t, parsedFrame.Hands[0].Z(), 76.3394)
